+++ app/scripts/components/search-examples.ts
@@ -7,7 +7,6 @@ import settings from "@/settings"
 import { SearchExample } from "@/settings/app-settings.types"
 import { HashParams } from "@/urlparams"
 import { CqpSearchEvent } from "@/statemachine/types"
-import { SearchesService } from "@/services/searches"
 
 export default angular.module("korpApp").component("searchExamples", {
     template: html`
@@ -27,8 +26,7 @@ export default angular.module("korpApp").component("searchExamples", {
     controller: [
         "$scope",
         "$location",
-        "searches",
-        function ($scope: SearchExamplesScope, $location: LocationService, searches: SearchesService) {
+        function ($scope: SearchExamplesScope, $location: LocationService) {
             const $ctrl = this
 
             $scope.examples = undefined
@@ -48,8 +46,6 @@ export default angular.module("korpApp").component("searchExamples", {
                 }
                 // Do not use `$location.search(params)` because it will remove existing params (like `corpus`)
                 Object.keys(params).forEach((key: keyof HashParams) => $location.search(key, params[key]))
-                // Let new values be synced to store before triggering search
-                setTimeout(() => searches.doSearch())
             }
         },
     ],